Mikkel
mikkels ting forsvandt, vi tog det tilbage

+++ spawnFruit.pde
@@ -2,8 +2,8 @@ boolean spawnNew = true;
 boolean availableHead = false;
 boolean availableTail = false;
 boolean available = false;
-float[] nFruitX = new float[cells], nFruitY = new float[cells];
-float fruitX, fruitY;
+int[] nFruitX = new int[cells], nFruitY = new int[cells];
+int fruitX, fruitY;
 
 void spawnFruit() {
   if (spawnNew == true) { //skal ny frugt spawne?
